requirements.txt
asgiref==3.5.0
Django==4.0.2
django-autoslug==1.9.8
django-ckeditor==6.2.0
django-js-asset==2.0.0
django-otp==1.1.3
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
